api/requirements.txt
# Framework
Flask==2.2

# For protection from CSRF attacks
Flask-WTF==1.0.1

# For ORM in Postgress
psycopg2-binary==2.8.4
Flask-SQLAlchemy==3.0.2

# For redis
redis==4.3.4

# For S3
boto3==1.24.84

# For dealing with the database
flask-migrate==3.1.0

# For authentication and authorization
bcrypt==4.0.0
PyJWT==1.7.1

# For environment variables
python-dotenv==0.19.0

# For testing
pytest==7.1.3
flask-testing==0.8.1
pytest-cov==3.0.0

